+++ src/components/ElasticsearchInstance/InstancesTable.vue
@@ -6,7 +6,7 @@
           <new-instance/>
         </v-col>
         <v-col>
-          <div class="float-right d-inline-block">
+          <div class="float-right">
             <v-text-field v-model="filter"
                           :label="$t('defaults.filter.label')"
                           append-icon="mdi-magnify"
@@ -23,12 +23,12 @@
     <v-data-table id="elasticsearch-clusters"
                   :footer-props="{itemsPerPageOptions: DEFAULT_ITEMS_PER_PAGE, showFirstLastPage: true}"
                   :headers="headers"
-                  :items="instances"
+                  :items="indexedInstances"
                   :search="filter"
                   dense>
-      <template v-slot:item="{ item, index }">
+      <template v-slot:item="{ item }">
         <tr :title="$t('elasticsearch_instance.instances_table.row.title', {uri: item.uri})" class="tr--clickable"
-            @click="setActiveInstanceIdx(index)">
+            @click="switchCluster(item.index)">
           <td class="pt-1">
             <div :title="$t('elasticsearch_instance.instances_table.row.cluster_health.title', {status: item.status})"
                  class="d-inline-block">
@@ -39,7 +39,7 @@
 
             <div class="d-inline-block text-truncate" style="max-width: 300px;">
               {{ item.name }}
-              <v-chip v-if="index === activeInstanceIdx" class="mx-1" color="success" small>active</v-chip>
+              <v-chip v-if="item.index === activeInstanceIdx" class="mx-1" color="success" small>active</v-chip>
             </div>
           </td>
           <td class="pt-1">
@@ -53,12 +53,12 @@
             {{ item.version }}
           </td>
           <td>
-            <rename-instance :cluster-idx="index" :cluster-name="item.name" :cluster-uri="item.uri"/>
+            <rename-instance :cluster-idx="item.index" :cluster-name="item.name" :cluster-uri="item.uri"/>
 
-            <v-btn :id="`remove-instance-${index}`" class="ml-1" icon
+            <v-btn :id="`remove-instance-${item.index}`" class="ml-1" icon
                    small
                    :title="$t('elasticsearch_instance.instances_table.row.remove_cluster.title')"
-                   @click.stop="removeInstance(index)">
+                   @click.stop="removeInstance(item.index)">
               <v-icon small>mdi-delete</v-icon>
             </v-btn>
           </td>
@@ -72,11 +72,14 @@
   import store from '@/store'
   import { vuexAccessors } from '@/helpers/store'
   import { BASE_URI, DEFAULT_ITEMS_PER_PAGE } from '@/consts'
-  import { ref } from '@vue/composition-api'
+  import { computed, ref } from 'vue'
   import i18n from '@/i18n'
   import RenameInstance from '@/components/ElasticsearchInstance/RenameInstance'
   import NewInstance from '@/components/ElasticsearchInstance/NewInstance'
   import CopyButton from '@/components/shared/CopyButton'
+  import { reloadHomePage } from '@/helpers'
+  import { askConfirm } from '@/services/tauri/dialogs'
+  import { useRouter } from '@/helpers/composition'
 
   export default {
     name: 'instances-table',
@@ -88,21 +91,24 @@
     setup () {
       const { activeInstanceIdx, instances } = vuexAccessors('connection', ['activeInstanceIdx', 'instances'])
 
-      const setActiveInstanceIdx = index => {
-        store.commit('connection/setActiveInstanceIdx', index)
-        window.location.replace(BASE_URI)
-      }
+      const indexedInstances = computed(() => {
+        return [...instances.value].map((instance, i) => Object.assign({}, instance, { index: i }))
+      })
+
+      const { router } = useRouter()
+      const switchCluster = index => reloadHomePage(router, index.toString())
 
       const removeInstance = index => {
-        if (confirm(i18n.t('elasticsearch_instance.instances_table.row.remove_cluster.confirm', {
-          name: instances.value[index].name,
-          uri: instances.value[index].uri
-        }))) {
-          let reload
-          if (index === activeInstanceIdx.value) reload = true
-          store.commit('connection/removeInstance', index)
-          if (reload) window.location.replace(BASE_URI)
-        }
+        askConfirm(i18n.t('elasticsearch_instance.instances_table.row.remove_cluster.confirm', {
+          name: instances.value[index].name, uri: instances.value[index].uri
+        })).then(confirmed => {
+          if (confirmed) {
+            let reset
+            if (index === activeInstanceIdx.value) reset = true
+            store.commit('connection/removeInstance', index)
+            if (reset) window.location.replace(BASE_URI)
+          }
+        })
       }
 
       const filter = ref('')
@@ -115,8 +121,8 @@
 
       return {
         activeInstanceIdx,
-        instances,
-        setActiveInstanceIdx,
+        indexedInstances,
+        switchCluster,
         removeInstance,
         filter,
         headers,