+++ src/components/ElasticsearchInstance/NewInstance.vue
@@ -132,12 +132,14 @@
 <script>
   import store from '@/store'
   import { vuexAccessors } from '@/helpers/store'
-  import { computed, ref } from '@vue/composition-api'
+  import { computed, ref } from 'vue'
   import { useTestConnection } from '@/mixins/TestConnection'
-  import { BASE_URI, SHOW_CORS_HINT } from '@/consts'
+  import { SHOW_CORS_HINT } from '@/consts'
   import Configure from '@/components/Setup/Configure'
   import SslHint from '@/components/shared/SslHint'
   import AuthorizationHeaderHint from '@/components/shared/AuthorizationHeaderHint'
+  import { reloadHomePage } from '@/helpers'
+  import { useRouter } from '@/helpers/composition'
 
   export default {
     name: 'new-instance',
@@ -170,12 +172,14 @@
         connect
       } = useTestConnection()
 
+      const { router } = useRouter()
       const connectCluster = () => {
         connect()
           .then(() => {
             if (!testState.value.connectError) {
-              store.commit('connection/setActiveInstanceIdx', instances.value.length - 1)
-              window.location.replace(BASE_URI)
+              const newId = instances.value.length - 1
+              store.commit('connection/setActiveInstanceIdx', newId)
+              reloadHomePage(router, newId)
             }
           })
       }