+++ src/store/modules/connection.js
@@ -1,9 +1,7 @@
-import { DEFAULT_ELASTICSEARCH_HOST } from '@/consts'
-
 export const connection = {
   namespaced: true,
   state: {
-    activeInstanceIdx: 0,
+    activeInstanceIdx: null,
     instances: [],
     elasticsearchAdapter: null
   },
@@ -20,13 +18,12 @@ export const connection = {
     },
     removeInstance (state, index) {
       state.instances.splice(index, 1)
-      if (index === state.activeInstanceIdx) {
+      if (state.instances.length === 0) {
+        state.activeInstanceIdx = null
+      } else if (index === state.activeInstanceIdx) {
         state.activeInstanceIdx = 0
       } else if (index < state.activeInstanceIdx) {
         state.activeInstanceIdx = state.activeInstanceIdx - 1
-      }
-      if (state.instances.length === 0) {
-        state.instances.push(Object.assign({}, DEFAULT_ELASTICSEARCH_HOST))
       }
     },
     setActiveInstanceIdx (state, index) {